--- packages/storybook/src/clinical/organisms/document-reference-timeline/document-timeline.fixtures.ts
@@ -1,321 +1,0 @@
-import { DocumentReference, DocumentReferenceContinuation, DocumentReferenceStatusCode } from '@ltht-react/types'
-
-const DocumentReference6: DocumentReference = {
-  id: 'c00dd3a5-e78a-4760-baa4-2f57edef4df8',
-  created: {
-    value: '2022-07-06T11:40:28.023+00:00',
-  },
-  description: 'Clinical Note (Active)',
-  text: {
-    text: 'Clinical Note (Active)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '618a16a5-a2ee-498f-8f80-5046e934871c',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'c00dd3a5-e78a-4760-baa4-2f57edef4df8',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReference5: DocumentReference = {
-  id: '8db4c96d-a443-46eb-a165-c5f28f87a0b6',
-  created: {
-    value: '2022-06-22T09:27:57.407+00:00',
-  },
-  description: 'Clinical Note (Active)',
-  text: {
-    text: 'Clinical Note (Active)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'ccff6894-20eb-4c57-8e79-e837498e43b3',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '8db4c96d-a443-46eb-a165-c5f28f87a0b6',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReference4: DocumentReference = {
-  id: '46263f65-d5bd-4197-89b0-f2e3c0b33665',
-  created: {
-    value: '2022-01-07T12:01:21.853+00:00',
-  },
-  description: 'Clinical Note (Deleted)',
-  text: {
-    text: 'Clinical Note (Deleted)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '941eda8f-2dbf-4a50-a0bb-d5776d1b814f',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '46263f65-d5bd-4197-89b0-f2e3c0b33665',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReference3: DocumentReference = {
-  id: 'bdfe6034-bb1b-4a02-9b03-fbdd8db2147a',
-  created: {
-    value: '2021-12-16T10:58:14.03+00:00',
-  },
-  description: 'Clinical Note (Deleted)',
-  text: {
-    text: 'Clinical Note (Deleted)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '6cde4dff-09f2-4f15-9477-f6b42c2adbe3',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'bdfe6034-bb1b-4a02-9b03-fbdd8db2147a',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReference2: DocumentReference = {
-  id: '8d4119a4-91c3-4e97-93cc-3c7ada7a4a5d',
-  created: {
-    value: '2021-12-15T14:20:20.55+00:00',
-  },
-  description: 'Clinical Note (Deleted)',
-  text: {
-    text: 'Clinical Note (Deleted)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'a56e3062-714c-404d-9042-79a12d90b192',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '8d4119a4-91c3-4e97-93cc-3c7ada7a4a5d',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReference1: DocumentReference = {
-  id: 'dc560990-e2bb-4aef-a1ab-9c1402ca75ce',
-  created: {
-    value: '2021-11-04T10:44:24.143+00:00',
-  },
-  description: 'Clinical Note (Deleted)',
-  text: {
-    text: 'Clinical Note (Deleted)',
-    div: '',
-  },
-  identifier: [
-    {
-      system: 'EventTypeName',
-      value: 'CdsRootActivated',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'clinical-note',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: '52c4d808-8984-43fe-b820-b2cfaaedaa46',
-    },
-    {
-      system: 'https://leedsth.nhs.uk/cds/template-name',
-      value: 'dc560990-e2bb-4aef-a1ab-9c1402ca75ce',
-    },
-  ],
-  metadata: {
-    extension: undefined,
-    versionId: undefined,
-    requestedWhen: '',
-    lastUpdated: undefined,
-    dataSources: [],
-    security: undefined,
-    isRedacted: false,
-    tag: undefined,
-  },
-  status: DocumentReferenceStatusCode.Current,
-  type: {
-    extension: undefined,
-    coding: undefined,
-    text: undefined,
-  },
-  indexed: {
-    value: undefined,
-    kind: undefined,
-  },
-  content: [],
-}
-
-const DocumentReferences: DocumentReferenceContinuation = {
-  resources: [
-    DocumentReference6,
-    DocumentReference5,
-    DocumentReference4,
-    DocumentReference3,
-    DocumentReference2,
-    DocumentReference1,
-  ],
-  selfCursorToken: '',
-}
-
-export default DocumentReferences
